Mailsender/requirements.txt
flask
mailjet-rest
gunicorn
python-dotenv
